[bugfix] clean up pid/tid parsing
Cleanup ProcessThreadId::parse so it will properly fail when stroul fails to
parse an integer value. It was relying on errno to detect stroul failure, but
errno is only guaranteed set on over/underflow errors so this check is
insufficient. Now it also compares the output and input string positions to see
if any chars were actually consumed.

Also delete the sketchy CHECK_AND_RESET macro and only update the member pid and
tid values on success so the macro isn't needed.